Operation brake-check the Hop-ons:
Problem:
SceneView.getSpriteSheet doesn't cache loaded resources, and hits resource load for each invocation.  This function is called when validating that a tile sprite is ready for rendering, and must be quick.

Solution:
Don't be an idiot and cache results on the view.  Object lookup is sufficiently fast.

Result:
getSpriteSheet gets road-rash as it's knocked off the top offender's stair-car.