Update StateMachine to allow UI blocking of transitions... COOL!
Example usage: The state machine does not know about sleeping after an attack to allow time for the UI to display a popup with damage text and wait for it to go away.  The UI listens to the combat:attack event, and then sets machine.uiBlocked = true.  It shows a message displaying the attack statistics, and then fades out after a timeout.  When it fades out it simply flips machine.uiBlocked to false, and the machine transitions to the next players turn.